requirements.txt
fastapi
uvicorn
pymongo
google-genai
certifi
python-consul
python-dotenv